+++ changelog.md
@@ -7,6 +7,7 @@ Core:
 - Allow to leave workspace
 - Allow to kick employee
 - Browser notifications
+- Allow to create employee
 
 HR: